Update to .NET 9 (#532)
* Update to .NET 9

Update eShop to use built-in OpenAPI support (#360)

* Update eShop to use built-in OpenAPI support

* Pass version names explicitly

Update dependencies

Downgrade EF Core, upgrade SDK, freeze workloads

Upgrade to working efcore and rename rollback file

Rearrange items in Directory.Packages.props

* Update Microsoft.Extensions.Http.Resilience dependency

* Remove unneeded OpenAPI workarounds

* Address feedback
