+++ global.json
@@ -1,6 +1,6 @@
 {
   "sdk": {
-    "version": "8.0.400",
+    "version": "9.0.100-rc.1.24452.12",
     "rollForward": "latestFeature",
     "allowPrerelease": true
   }
