+++ src/Ordering.API/GlobalUsings.cs
@@ -37,4 +37,3 @@ global using Microsoft.Extensions.Options;
 global using Polly;
 global using Polly.Retry;
 global using eShop.ServiceDefaults;
-global using Swashbuckle.AspNetCore.SwaggerGen;
