+++ src/eShop.ServiceDefaults/OpenApi.Extensions.cs
@@ -1,11 +1,12 @@
 ﻿using Asp.Versioning;
+using Asp.Versioning.ApiExplorer;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Routing;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
-using Microsoft.Extensions.Options;
-using Swashbuckle.AspNetCore.SwaggerGen;
+using Scalar.AspNetCore;
 
 namespace eShop.ServiceDefaults;
 
@@ -21,45 +22,12 @@ public static partial class Extensions
             return app;
         }
 
-        app.UseSwagger();
+        app.MapOpenApi();
 
         if (app.Environment.IsDevelopment())
         {
-            app.UseSwaggerUI(setup =>
-            {
-                /// {
-                ///   "OpenApi": {
-                ///     "Endpoint: {
-                ///         "Name": 
-                ///     },
-                ///     "Auth": {
-                ///         "ClientId": ..,
-                ///         "AppName": ..
-                ///     }
-                ///   }
-                /// }
-
-                var pathBase = configuration["PATH_BASE"] ?? string.Empty;
-                var authSection = openApiSection.GetSection("Auth");
-                var endpointSection = openApiSection.GetRequiredSection("Endpoint");
-
-                foreach (var description in app.DescribeApiVersions())
-                {
-                    var name = description.GroupName;
-                    var url = endpointSection["Url"] ?? $"{pathBase}/swagger/{name}/swagger.json";
-
-                    setup.SwaggerEndpoint(url, name);
-                }
-
-                if (authSection.Exists())
-                {
-                    setup.OAuthClientId(authSection.GetRequiredValue("ClientId"));
-                    setup.OAuthAppName(authSection.GetRequiredValue("AppName"));
-                }
-            });
-
-            // Add a redirect from the root of the app to the swagger endpoint
-            app.MapGet("/", () => Results.Redirect("/swagger")).ExcludeFromDescription();
+            app.MapScalarApiReference();
+            app.MapGet("/", () => Results.Redirect("/scalar/v1")).ExcludeFromDescription();
         }
 
         return app;
@@ -69,24 +37,35 @@ public static partial class Extensions
         this IHostApplicationBuilder builder,
         IApiVersioningBuilder? apiVersioning = default)
     {
-        var services = builder.Services;
-        var configuration = builder.Configuration;
-        var openApi = configuration.GetSection("OpenApi");
+        var openApi = builder.Configuration.GetSection("OpenApi");
+        var identitySection = builder.Configuration.GetSection("Identity");
+
+        var scopes = identitySection.Exists()
+            ? identitySection.GetRequiredSection("Scopes").GetChildren().ToDictionary(p => p.Key, p => p.Value)
+            : new Dictionary<string, string?>();
+
 
         if (!openApi.Exists())
         {
             return builder;
         }
 
-        services.AddEndpointsApiExplorer();
-
         if (apiVersioning is not null)
         {
             // the default format will just be ApiVersion.ToString(); for example, 1.0.
             // this will format the version as "'v'major[.minor][-status]"
-            apiVersioning.AddApiExplorer(options => options.GroupNameFormat = "'v'VVV");
-            services.AddTransient<IConfigureOptions<SwaggerGenOptions>, ConfigureSwaggerOptions>();
-            services.AddSwaggerGen(options => options.OperationFilter<OpenApiDefaultValues>());
+            var versioned = apiVersioning.AddApiExplorer(options => options.GroupNameFormat = "'v'VVV");
+            string[] versions = ["v1"];
+            foreach (var description in versions)
+            {
+                builder.Services.AddOpenApi(description, options =>
+                {
+                    options.ApplyApiVersionInfo(openApi.GetRequiredValue("Document:Title"), openApi.GetRequiredValue("Document:Description"));
+                    options.ApplyAuthorizationChecks([.. scopes.Keys]);
+                    options.ApplySecuritySchemeDefinitions();
+                    options.ApplyOperationDeprecatedStatus();
+                });
+            }
         }
 
         return builder;
