--- src/eShop.ServiceDefaults/OpenApiDefaultValues.cs
@@ -1,58 +1,0 @@
-﻿using Microsoft.AspNetCore.Mvc.ApiExplorer;
-using Microsoft.AspNetCore.Mvc.ModelBinding;
-using Microsoft.OpenApi.Models;
-using Swashbuckle.AspNetCore.SwaggerGen;
-using System.Text.Json;
-
-namespace eShop.ServiceDefaults;
-
-internal sealed class OpenApiDefaultValues : IOperationFilter
-{
-    public void Apply(OpenApiOperation operation, OperationFilterContext context)
-    {
-        var apiDescription = context.ApiDescription;
-
-        operation.Deprecated |= apiDescription.IsDeprecated();
-
-        // remove any assumed media types not present in the api description
-        foreach (var responseType in context.ApiDescription.SupportedResponseTypes)
-        {
-            var responseKey = responseType.IsDefaultResponse ? "default" : responseType.StatusCode.ToString();
-            var response = operation.Responses[responseKey];
-
-            foreach (var contentType in response.Content.Keys)
-            {
-                if (!responseType.ApiResponseFormats.Any(x => x.MediaType == contentType))
-                {
-                    response.Content.Remove(contentType);
-                }
-            }
-        }
-
-        if (operation.Parameters == null)
-        {
-            return;
-        }
-
-        // fix-up parameters with additional information from the api explorer that might
-        // not have otherwise been used. this will most often happen for api version parameters
-        // which are dynamically added, have no endpoint signature info, nor any xml comments.
-        foreach (var parameter in operation.Parameters)
-        {
-            var description = apiDescription.ParameterDescriptions.First(p => p.Name == parameter.Name);
-
-            parameter.Description ??= description.ModelMetadata?.Description;
-
-            if (parameter.Schema.Default == null &&
-                description.DefaultValue != null &&
-                description.DefaultValue is not DBNull &&
-                description.ModelMetadata is ModelMetadata modelMetadata)
-            {
-                var json = JsonSerializer.Serialize(description.DefaultValue, modelMetadata.ModelType);
-                parameter.Schema.Default = OpenApiAnyFactory.CreateFromJson(json);
-            }
-
-            parameter.Required |= description.IsRequired;
-        }
-    }
-}
